+++ tests/target/configs/spaces_around_ranges/true.rs
@@ -1,6 +1,0 @@
-// rustfmt-spaces_around_ranges: true
-// Spaces around ranges
-
-fn main() {
-    let lorem = 0 .. 10;
-}