+++ tests/target/macro_not_expr.rs
@@ -1,7 +1,0 @@
-macro_rules! test {
-    ($($t:tt)*) => {}
-}
-
-fn main() {
-    test!( a : B => c d );
-}